correcao da tello_zune, criacao de scripts exemplos
