+++ README.md
@@ -1,2 +1,4 @@
-# tello_drone
-Projetos utilizando o drone DJI Tello, utilizando Python.
+# tello_zune
+Biblioteca utilizada pela equipe Zune Drones para controlar o drone DJI Tello. Na pasta examples tem alguns exemplos simples de utilização do módulo tello_zune, que também está presente na pasta.
+
+Para utiizar o módulo tello_zune é necessário instalar o FFMPEG.
