+++ examples/detectionfaces_tello.py
@@ -1,23 +1,19 @@
 import cv2
-from djitellopy import Tello
+from tello_zune import TelloZune
 
 # inicializa e conecta com o Tello
-tello = Tello()
-tello.connect()
+tello = TelloZune()
+tello.start_tello()
 # cria o objeto face_detector para fazer a deteccao de faces
 face_detector = cv2.CascadeClassifier('cascades\haarcascade_frontalface_default.xml')
-# inicializa a tranmissao de video
-tello.streamon()
 
 while True:
     # captura cada frame do objeto frame do tello
-    frame = tello.get_frame_read().frame
-    # converte cada frame de RGB para BGR
-    frame = cv2.cvtColor(frame, cv2.COLOR_RGB2BGR)
+    frame = tello.get_frame()
     # converte a imagem de frame BGR para escala de cinza para a deteccao
     image_gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
     # faz a deteccao a cada frame
-    detections = face_detector.detectMultiScale(image_gray, minSize=(130, 130), minNeighbors=9, maxSize=(290, 290))
+    detections = face_detector.detectMultiScale(image_gray, minSize=(130, 130), minNeighbors=9, maxSize=(290, 290), scaleFactor=1.05)
 
     #exibe o numero de rostos detectados
     if len(detections) >= 1:
@@ -35,5 +31,5 @@ while True:
     if cv2.waitKey(1) & 0xFF == ord('q'):
         break
 # desliga a transmissao do video
-tello.streamoff()
+tello.end_tello()
 cv2.destroyAllWindows()
